Update README.md
A simple HTML, CSS, and JavaScript calculator project is hosted on this GitHub repository. The calculator has an easy-to-use interface and offers all the necessary arithmetic functions. This repository is a great place to start if you're new to web development or just looking for a simple calculator for your projects. You are welcome to investigate the code, make changes, or incorporate it into your web apps. The project places a strong emphasis on simplicity, making it simple to use and adapt. Have fun with coding!